% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_csf.R
\name{add_ground_csf}
\alias{add_ground_csf}
\title{Add ground as class 3 using CSF after FuelDeep3D prediction}
\usage{
add_ground_csf(in_las, out_las, csf_args = list())
}
\arguments{
\item{in_las}{LAS/LAZ with FuelDeep3D predictions in Classification (0,1,2)}

\item{out_las}{Output LAS/LAZ with ground set to class 3}

\item{csf_args}{list of arguments passed to lidR::csf()}
}
\description{
Add ground as class 3 using CSF after FuelDeep3D prediction
}
