% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_las}
\alias{evaluate_las}
\title{Evaluate Predicted vs True LAS Labels}
\usage{
evaluate_las(pred, truth)
}
\arguments{
\item{pred}{LAS with predicted Classification}

\item{truth}{LAS with true labels (column "label")}
}
\value{
list of confusion matrix, accuracy, precision, recall, F1
}
\description{
Evaluate Predicted vs True LAS Labels
}
