% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_two_las}
\alias{evaluate_two_las}
\title{Evaluate two LAS files (truth vs prediction)}
\usage{
evaluate_two_las(
  truth_las,
  pred_las,
  truth_col = "label",
  pred_col = "Classification"
)
}
\arguments{
\item{truth_las}{A \code{lidR::LAS} object containing ground-truth labels.}

\item{pred_las}{A \code{lidR::LAS} object containing predicted labels.}

\item{truth_col}{Character. Name of truth label field in \code{truth_las} (default = "label").}

\item{pred_col}{Character. Name of prediction field in \code{pred_las} (default = "Classification").}
}
\value{
A list with confusion matrix, accuracy, precision, recall, and F1.
}
\description{
Computes confusion matrix, accuracy, and per-class precision/recall/F1
using two separate LAS objects that are point-wise aligned (same number of points).
}
\examples{
\donttest{
library(lidR)
truth <- readLAS(system.file("extdata", "las", "tree2.laz", package = "FuelDeep3D"))
pred  <- readLAS(system.file("extdata", "las", "tree2.laz", package = "FuelDeep3D"))
if (!lidR::is.empty(truth) && !lidR::is.empty(pred)) {
  truth@data$label <- as.integer(truth@data$Classification)
  res <- evaluate_two_las(truth, pred, truth_col = "label", pred_col = "Classification")
  res$accuracy
}
}

}
