% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_las_3d}
\alias{plot_las_3d}
\title{3D Visualization of LAS Using RGL}
\usage{
plot_las_3d(las, palette = palette_veg(), size = 2, title = "LAS 3D View")
}
\arguments{
\item{las}{LAS object}

\item{palette}{Color palette for classes}

\item{size}{Point size}

\item{title}{Plot title}
}
\description{
3D Visualization of LAS Using RGL
}
