% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_confusion_matrix}
\alias{print_confusion_matrix}
\title{Pretty-print confusion matrix}
\usage{
print_confusion_matrix(cm)
}
\arguments{
\item{cm}{Confusion matrix (table)}
}
\description{
Pretty-print confusion matrix
}
