% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_setup.r
\name{py_setup}
\alias{py_setup}
\title{Create/use a venv and install Python deps listed in extdata/python/requirements.txt}
\usage{
py_setup(envdir = NULL)
}
\description{
Create/use a venv and install Python deps listed in extdata/python/requirements.txt
}
