% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict}
\alias{predict}
\title{Run prediction on cfg$las_path and write output LAS
If cfg$num_classes == 4, add ground using CSF automatically.}
\usage{
predict(
  cfg,
  mode = c("overwrite", "extra"),
  setup_env = FALSE,
  csf_args = list()
)
}
\arguments{
\item{cfg}{list from config()}

\item{mode}{"overwrite" or "extra"}

\item{setup_env}{logical; FALSE if you already have Python env}

\item{csf_args}{list passed to add_ground_csf() if num_classes == 4}
}
\value{
path to output LAS
}
\description{
Run prediction on cfg$las_path and write output LAS
If cfg$num_classes == 4, add ground using CSF automatically.
}
