% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict}
\alias{predict}
\title{Run prediction on cfg$las_path and write output LAS}
\usage{
predict(cfg, mode = c("overwrite", "extra"), setup_env = FALSE)
}
\arguments{
\item{cfg}{list from config()}

\item{mode}{"overwrite" or "extra"}

\item{setup_env}{logical; FALSE if you already have Python env}
}
\value{
output LAS path (character)
}
\description{
Run prediction on cfg$las_path and write output LAS
}
