% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_confusion_matrix}
\alias{print_confusion_matrix}
\title{Pretty-print confusion matrix}
\usage{
print_confusion_matrix(cm)
}
\arguments{
\item{cm}{Confusion matrix (table or matrix).}
}
\value{
Invisibly returns the printed data.frame.
}
\description{
Converts a confusion matrix to a data.frame and prints it.
}
\examples{
cm <- table(True = c(0, 0, 1, 1), Pred = c(0, 1, 1, 1))
print_confusion_matrix(cm)

}
