% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_metrics_table}
\alias{print_metrics_table}
\title{Pretty-print evaluation metrics table with summary row}
\usage{
print_metrics_table(results)
}
\arguments{
\item{results}{List returned by evaluate_single_las() or evaluate_two_las()}
}
\description{
Pretty-print evaluation metrics table with summary row
}
