% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{train}
\alias{train}
\title{Train the FuelDeep3D model (auto-preprocess if NPZ tiles are missing)}
\usage{
train(cfg, setup_env = FALSE)
}
\arguments{
\item{cfg}{list from config()}

\item{setup_env}{logical; set TRUE to create/use the package venv}
}
\value{
list(best_oa, best_epoch, ckpt_path)
}
\description{
Train the FuelDeep3D model (auto-preprocess if NPZ tiles are missing)
}
